# cap-sistema/docs/openapi/programas.yaml
openapi: 3.0.0
paths:
...
  /v1/programas/hipertension:
    get:
      operationId: HipertensionController_listar
      parameters:
        - name: estado
          required: false
          in: query
          schema:
            enum:
              - ACTIVO
              - EGRESADO
              - ABANDONO
              - FALLECIDO
              - TRASLADADO
            type: string
        - name: comunidadId
          required: false
          in: query
          schema:
            format: uuid
            type: string
        - name: pagina
          required: false
          in: query
          description: Empieza en 1. Por defecto 1.
          schema:
            example: 1
            type: number
        - name: tamano
          required: false
          in: query
          description: Por defecto 25. El servidor recorta cualquier valor mayor a 100.
          schema:
            example: 25
            type: number
      responses:
        "200":
          description: "El ultimo control viene incrustado: evita el N+1 de esta pantalla."
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginaDto"
                  - type: object
                    required:
                      - datos
                    properties:
                      datos:
                        type: array
                        items:
                          $ref: "#/components/schemas/ProgramaHipertensionResumenDto"
        "400":
          description: La informacion enviada no es valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "401":
          description: Falta el token, expiro o no es valido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "403":
          description: El rol de la cuenta no tiene permiso sobre este recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "500":
          description: Error inesperado. El mensaje real queda en los logs, no se expone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
      security: &a2
        - bearer: []
      summary: Inscripciones, con su ultimo control
      tags: &a3
        - programa-hipertension
    post:
      operationId: HipertensionController_inscribir
      parameters:
        - name: authorization
          required: true
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InscribirHipertensionDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramaHipertensionDto"
        "400":
          description: La informacion enviada no es valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "401":
          description: Falta el token, expiro o no es valido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "403":
          description: El rol de la cuenta no tiene permiso sobre este recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "500":
          description: Error inesperado. El mensaje real queda en los logs, no se expone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
      security: *a2
      summary: Inscribe a un paciente en el programa
      tags: *a3
  /v1/programas/hipertension/atrasados:
    get:
      operationId: HipertensionController_atrasados
      parameters:
        - name: pagina
          required: false
          in: query
          description: Empieza en 1. Por defecto 1.
          schema:
            example: 1
            type: number
        - name: tamano
          required: false
          in: query
          description: Por defecto 25. El servidor recorta cualquier valor mayor a 100.
          schema:
            example: 25
            type: number
      responses:
        "200":
          description: Ordenados del mas atrasado al menos.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginaDto"
                  - type: object
                    required:
                      - datos
                    properties:
                      datos:
                        type: array
                        items:
                          $ref: "#/components/schemas/HipertensoAtrasadoDto"
        "400":
          description: La informacion enviada no es valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "401":
          description: Falta el token, expiro o no es valido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "403":
          description: El rol de la cuenta no tiene permiso sobre este recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "500":
          description: Error inesperado. El mensaje real queda en los logs, no se expone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
      security: *a2
      summary: Pacientes que ya pasaron su fecha de proximo control
      tags: *a3
  /v1/programas/hipertension/{id}:
    get:
      operationId: HipertensionController_obtener
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramaHipertensionDto"
        "400":
          description: La informacion enviada no es valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "401":
          description: Falta el token, expiro o no es valido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "403":
          description: El rol de la cuenta no tiene permiso sobre este recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "404":
          description: El recurso solicitado no existe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "500":
          description: Error inesperado. El mensaje real queda en los logs, no se expone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
      security: *a2
      summary: Inscripcion por su identificador
      tags: *a3
  /v1/programas/hipertension/{id}/controles:
    get:
      operationId: HipertensionController_listarControles
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: pagina
          required: false
          in: query
          description: Empieza en 1. Por defecto 1.
          schema:
            example: 1
            type: number
        - name: tamano
          required: false
          in: query
          description: Por defecto 25. El servidor recorta cualquier valor mayor a 100.
          schema:
            example: 25
            type: number
      responses:
        "200":
          description: Listado paginado.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginaDto"
                  - type: object
                    required:
                      - datos
                    properties:
                      datos:
                        type: array
                        items:
                          $ref: "#/components/schemas/ControlHipertensionDto"
        "400":
          description: La informacion enviada no es valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "401":
          description: Falta el token, expiro o no es valido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "403":
          description: El rol de la cuenta no tiene permiso sobre este recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "404":
          description: El recurso solicitado no existe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "500":
          description: Error inesperado. El mensaje real queda en los logs, no se expone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
      security: *a2
      summary: Controles del programa, lo mas reciente primero
      tags: *a3
    post:
      description: La clasificacion y la fecha del proximo control las calcula el
        sistema; no se teclean.
      operationId: HipertensionController_registrarControl
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrarControlHipertensionDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlHipertensionDto"
        "400":
          description: La informacion enviada no es valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "401":
          description: Falta el token, expiro o no es valido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "403":
          description: El rol de la cuenta no tiene permiso sobre este recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "404":
          description: El recurso solicitado no existe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "500":
          description: Error inesperado. El mensaje real queda en los logs, no se expone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
      security: *a2
      summary: Registra un control
      tags: *a3
  /v1/programas/hipertension/{id}/egreso:
    patch:
      operationId: HipertensionController_egresar
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramaHipertensionDto"
        "400":
          description: La informacion enviada no es valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "401":
          description: Falta el token, expiro o no es valido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "403":
          description: El rol de la cuenta no tiene permiso sobre este recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "404":
          description: El recurso solicitado no existe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "500":
          description: Error inesperado. El mensaje real queda en los logs, no se expone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
      security: *a2
      summary: Cierra la inscripcion en el programa
      tags: *a3
...
info:
  title: Servicio programas - CAP Purulha
  description: >-
    Contrato del servicio programas.


    Generado desde el codigo. No editar a mano: los cambios se pierden en la
    siguiente exportacion. Fuente de verdad: arquitectura-cap-purulha.md
  version: "1.0"
  contact: {}
...
servers:
  - url: http://localhost:3003
    description: Desarrollo local
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    PaginaDto:
      type: object
      properties:
        pagina:
          type: number
          description: Pagina devuelta, empezando en 1.
          example: 1
        tamano:
          type: number
          description: Registros por pagina. El servidor recorta cualquier valor mayor al
            maximo.
          maximum: 100
          example: 25
        total:
          type: number
          description: Total de registros que cumplen el filtro.
          example: 137
        totalPaginas:
          type: number
          description: Total de paginas. Nunca es menor que 1.
          example: 6
      required:
        - pagina
        - tamano
        - total
        - totalPaginas
    UltimoControlHipertensionDto:
      type: object
      properties:
        fecha:
          format: date-time
          type: string
        sistolica:
          type: number
          example: 148
        diastolica:
          type: number
          example: 94
        clasificacion:
          type: string
          enum: &a7
            - NORMAL
            - ELEVADA
            - ESTADIO_1
            - ESTADIO_2
            - CRISIS
          description: Calculada a partir de las cifras, nunca tecleada.
        enMeta:
          type: boolean
          description: true si la lectura alcanza la meta acordada.
        proximoControl:
          type: string
          format: date-time
          nullable: true
      required:
        - fecha
        - sistolica
        - diastolica
        - clasificacion
        - enMeta
        - proximoControl
    ProgramaHipertensionResumenDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pacienteId:
          type: string
          format: uuid
        comunidadId:
          type: string
          format: uuid
        fechaIngreso:
          format: date-time
          type: string
        estado:
          type: string
          enum: &a6
            - ACTIVO
            - EGRESADO
            - ABANDONO
            - FALLECIDO
            - TRASLADADO
        metaSistolica:
          type: number
          example: 140
        metaDiastolica:
          type: number
          example: 90
        ultimoControl:
          nullable: true
          type: object
          allOf:
            - $ref: "#/components/schemas/UltimoControlHipertensionDto"
      required:
        - id
        - pacienteId
        - comunidadId
        - fechaIngreso
        - estado
        - metaSistolica
        - metaDiastolica
        - ultimoControl
    HipertensoAtrasadoDto:
      type: object
      properties:
        programaId:
          type: string
          format: uuid
        pacienteId:
          type: string
          format: uuid
        comunidadId:
          type: string
          format: uuid
        proximoControl:
          format: date-time
          type: string
        diasDeAtraso:
          type: number
          description: Dias transcurridos desde la fecha en que debio venir.
          example: 23
      required:
        - programaId
        - pacienteId
        - comunidadId
        - proximoControl
        - diasDeAtraso
    ProgramaHipertensionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pacienteId:
          type: string
          format: uuid
          description: Paciente del servicio usuarios. No es llave foranea.
        comunidadId:
          type: string
          format: uuid
        fechaIngreso:
          format: date-time
          type: string
        estado:
          type: string
          enum: *a6
        fechaEgreso:
          type: string
          format: date-time
          nullable: true
        motivoEgreso:
          type: string
          nullable: true
          maxLength: 200
        metaSistolica:
          type: number
          description: Meta acordada con el paciente, no un valor fijo del sistema.
          example: 140
        metaDiastolica:
          type: number
          example: 90
        inscritoPor:
          type: string
          format: uuid
        creadoEn:
          format: date-time
          type: string
      required:
        - id
        - pacienteId
        - comunidadId
        - fechaIngreso
        - estado
        - fechaEgreso
        - motivoEgreso
        - metaSistolica
        - metaDiastolica
        - inscritoPor
        - creadoEn
    InscribirHipertensionDto:
      type: object
      properties:
        pacienteId:
          type: string
          description: Identificador del paciente en el servicio de usuarios
        metaSistolica:
          type: number
          default: 140
          description: Meta de presion sistolica acordada
        metaDiastolica:
          type: number
          default: 90
      required:
        - pacienteId
    ControlHipertensionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        programaId:
          type: string
          format: uuid
        fecha:
          format: date-time
          type: string
        sistolica:
          type: number
          example: 148
        diastolica:
          type: number
          example: 94
        clasificacion:
          type: string
          enum: *a7
        enMeta:
          type: boolean
        pesoKg:
          type: string
          format: decimal
          nullable: true
          example: "72.50"
        adherencia:
          type: boolean
          nullable: true
          description: Si el paciente refiere tomar el tratamiento.
        proximoControl:
          type: string
          format: date-time
          nullable: true
          description: Calculado segun la clasificacion.
        registradoPor:
          type: string
          format: uuid
        observaciones:
          type: string
          nullable: true
          description: Descifradas. En la base son ilegibles.
      required:
        - id
        - programaId
        - fecha
        - sistolica
        - diastolica
        - clasificacion
        - enMeta
        - pesoKg
        - adherencia
        - proximoControl
        - registradoPor
        - observaciones
    RegistrarControlHipertensionDto:
      type: object
      properties:
        sistolica:
          type: number
          example: 148
          minimum: 50
          maximum: 300
        diastolica:
          type: number
          example: 94
          minimum: 30
          maximum: 200
        pesoKg:
          type: number
          minimum: 20
          maximum: 400
        adherencia:
          type: boolean
          description: Si el paciente refiere estar tomando el tratamiento
        observaciones:
          type: string
        fecha:
          format: date-time
          type: string
          description: Fecha real del control. Se usa al digitalizar papel.
      required:
        - sistolica
        - diastolica
...
    RespuestaErrorDto:
      type: object
      properties:
        codigo:
          type: string
          description: Codigo estable. El cliente decide que hacer segun este valor, no
            segun el mensaje.
          enum:
            - VALIDACION
            - NO_AUTENTICADO
            - SIN_PERMISO
            - NO_ENCONTRADO
            - CONFLICTO
            - DEMASIADAS_PETICIONES
            - ERROR_INTERNO
          example: VALIDACION
        mensaje:
          type: string
          description: Mensaje en espanol, apto para mostrarse tal cual al usuario.
          example: La informacion enviada no es valida.
        detalles:
          description: Detalle por campo. Presente sobre todo en errores de validacion.
          example:
            - dpi debe tener 13 digitos
          type: array
          items:
            type: string
        trazaId:
          type: string
          description: Identificador de correlacion. Es lo que se busca en los logs para
            esta peticion.
          example: 7f3a1c2e-9b4d-4e21-8a55-0c1d2e3f4a5b
        ruta:
          type: string
          example: /v1/pacientes
        fecha:
          type: string
          format: date-time
          example: 2026-08-26T15:04:05.000Z
      required:
        - codigo
        - mensaje
        - trazaId
        - ruta
        - fecha